Updated homotopy method, condition numbers, numerous examples in Example_x_y, black box solver in poly_rect_multipareig_homotopy.m

+++ find_duplicates.m
@@ -17,7 +17,7 @@ n = size(A,1);
 m = size(A,2);
 for j = 1:n
     elem = A(j,:);
-    if min(vecnorm(A([1:j-1 j+1:end],:) - elem,2,2))/(1+norm(elem)) < tol
+    if min(vecnorm(double(A([1:j-1 j+1:end],:) - elem),2,2))/(1+norm(elem)) < tol
         ind = [ind j];
     end
 end
